lint(coin-stellar): remove cyclic import
