+++ .changeset/gorgeous-grapes-scream.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-stellar": minor
+---
+
+lint: remove cyclic imports
