+++ libs/coin-modules/coin-stellar/.eslintrc.js
@@ -3,11 +3,14 @@ module.exports = {
     browser: true,
     es6: true,
   },
+  extends: ["plugin:import/typescript"],
+  plugins: ["import"],
   rules: {
     eqeqeq: ["error"],
     "no-console": ["error", { allow: ["warn", "error"] }],
     "@typescript-eslint/no-empty-function": "off",
     "@typescript-eslint/no-explicit-any": "warn",
+    "import/no-cycle": ["error"],
   },
   overrides: [
     {
